requirements.txt
discord.py
aiosqlite
discord.py-pagination